Se crea verificacion en el service de becas para la funcion de borrar y modificar becas para que no se puedan realizar durante el periodo de postulacion de la beca a modificar o eliminar

+++ backend/src/services/becas.service.js
@@ -73,13 +73,21 @@ async function updateBeca(id, bec) {
     try {
       const becFound = await Becas.findById(id);
       if (!becFound) return [null, "la beca no existe"];
-      const { nombre, requisitos, documentos, fecha_inicio, fecha_fin, monto, tipo_pago } = bec;
       const fecha_actual = Date.now();
+      /*Verificar que no se este modificando una beca durante el periodo de postulacion*/
+      const fechaf = new Date(becFound.fecha_fin);
+      const fechai = new Date(becFound.fecha_inicio);
+      if (fecha_actual > fechai && fecha_actual < fechaf){
+        return [null, "No se pueden modificar becas en periodo de postulacion"];
+      }
+      const { nombre, requisitos, documentos, fecha_inicio, fecha_fin, monto, tipo_pago } = bec;
+      /*Verificar que la fecha actual no este entre las fechas de postulacion*/
       const fechafin = new Date(bec.fecha_fin);
       const fechainicio = new Date(bec.fecha_inicio);
       if (fecha_actual > fechainicio && fecha_actual < fechafin){
-        return [null, "No se pueden modificar becas en periodo de postulacion"];
+        return [null, "No puedes estar entre el periodo de postulacion"];
       }
+
       const becUpdated = await Becas.findByIdAndUpdate(
         id,
         {
@@ -106,6 +114,14 @@ async function updateBeca(id, bec) {
  */
 async function BorrarBeca(id) {
     try {
+      const becas = await Becas.findById({ _id: id }) 
+      if (!becas) return [null, "La beca no existe"];
+      const fecha_actual = Date.now();
+      const fechaf = new Date(becas.fecha_fin);
+      const fechai = new Date(becas.fecha_inicio);
+      if (fecha_actual > fechai && fecha_actual < fechaf){
+        return [null, "No puedes borrar una beca en periodo de postulacion"];
+      }
       return await Becas.findByIdAndDelete(id);
     } catch (error) {
       handleError(error, "becas.service -> BorrarBeca");
